structs: word-sized pointers typed by their element (both access paths merged, validated strides folded); asserts seed via raw strings; naming bundle reads retail strings

+++ src/rel/main_rel/car/fn_1_8C51C.c
@@ -6,7 +6,7 @@ f32 fn_1_8C51C(u32 index) {
     u32 addr;
     u32 ptr;
 
-    base = lbl_1_bss_6D838;
+    base = (u32)lbl_1_bss_6D838;
     if (base != 0) {
         addr = base + index * 0x620;
     } else {
