+++ src/rel/main_rel/car/fn_1_8C6D4.c
@@ -6,7 +6,7 @@ u32 fn_1_8C6D4(u32 index) {
     u32 state;
 
     // Return the requested car's value from its state block.
-    base = lbl_1_bss_6D838;
+    base = (u32)lbl_1_bss_6D838;
     if (base != 0) {
         car = base + index * 0x620;
     } else {
